refactor(energy): the photovoltaic product, and solar.py is gone
The 733 lines left in solar.py were five subjects, and every one of its 53
remaining top-level symbols is placed:

  pv.py                   the chain for one array: Perez transposition, the
                          ASHRAE incidence modifier, Faiman module temperature,
                          PVWatts DC and inverter, over a 1 kWp reference
  terrain_irradiance.py   what the terrain does to it, in both directions: the
                          gain from slope and aspect, the loss to the pixel's
                          own horizon
  seasons.py              the month windows a seasonal figure is cut on, read
                          by two products in this slice and beside neither
  overlays.py             which palette a quantity gets and which domain it is
                          drawn on, including the two layers that must share one
  siting.py               slope limits and cover eligibility, the legend, the
                          class areas and the overlay

sidecar/solar.py is deleted.

THE ONLY EDGES OUT are terrain_irradiance -> pv, for the transposition it
applies per cell, and terrain_irradiance -> terra.sun.position, for
N_HORIZON_AZIMUTHS. SOLAR_CONSTANT went to pv beside its single reader rather
than to the sun service, which nothing in terra/sun reads, so pv imports
nothing from terra at all.

THE STRINGS. Deleting a module that the application cites on screen is a
user-visible change, and eight places named it:

  terra/energy/wind.py    six, one of them the data_quality flag that crosses
                          the JSON boundary and reaches the panel
  energy.py               two loss-waterfall rows, "definitional at
                          solar.PDC0_W" and the inverter oversizing note
  frontend               methodBrief.ts, the provenance line under the solar
                          terrain method; SolarSections.tsx, the season pair;
                          types.ts, the roughness note
  internal/research       two recorded sidecar payloads, wind_b.json and
                          energy_model_b.json, which exist so the exporter is
                          tested against the shape the sidecar actually emits;
                          and two comments in export_research.go

All corrected to the module that now holds the thing. The alternative was an
application citing a file that does not exist.

WHAT WENT WRONG ONCE. The first attempt at the call-site rewiring inserted the
new imports and then deleted the obsolete ones by their original line numbers,
which the insertions had already shifted. terra.actions imported cleanly and
action_solar_resource had simply lost its body. The deferred-import guard
failed on it; the second attempt rebuilds the line list from recorded edits
instead of mutating it twice.

tests/test_solar.py is repointed but keeps its name for one more commit: it now
exercises both sides, 24 tests on the service and 17 on the product, and
splitting it is a test reorganisation with no production change, which reads
better on its own.

Verification: pytest 511 passed, 8 skipped. gofmt, go vet and go test -race
clean, internal/research included, against the updated fixtures. A ping through
the entry point answers.
