+++ sidecar/tests/test_solar.py
@@ -5,7 +5,12 @@ from __future__ import annotations
 import numpy as np
 import pandas as pd
 
-import solar
+from terra.energy import (
+    overlays as overlays_mod,
+    terrain_irradiance as poa_mod,
+    pv as pv_mod,
+    seasons as seasons_mod,
+)
 from terra.sun import (
     position as sun_position,
     nasa_power as sun_power,
@@ -119,20 +124,20 @@ def test_reference_performance_ratio_is_below_the_modelled_one():
     cabling, so what it models runs high. The reference applied by default must
     sit below it, or the calibration would not be doing anything.
     """
-    assert solar.REFERENCE_PERFORMANCE_RATIO < 0.85
-    assert 0.7 < solar.REFERENCE_PERFORMANCE_RATIO < 0.9
+    assert pv_mod.REFERENCE_PERFORMANCE_RATIO < 0.85
+    assert 0.7 < pv_mod.REFERENCE_PERFORMANCE_RATIO < 0.9
 
 
 def test_modelled_performance_ratio_is_energy_over_reference_energy():
     p_ac = pd.Series([800.0, 800.0])
     poa = pd.Series([1000.0, 1000.0])
     # 1.6 kWh AC over 2.0 kWh at STC efficiency for a 1 kWp array.
-    assert abs(solar.modelled_performance_ratio(p_ac, poa) - 0.8) < 1e-9
+    assert abs(pv_mod.modelled_performance_ratio(p_ac, poa) - 0.8) < 1e-9
 
 
 def test_modelled_performance_ratio_handles_no_irradiance():
     assert np.isnan(
-        solar.modelled_performance_ratio(pd.Series([0.0]), pd.Series([0.0]))
+        pv_mod.modelled_performance_ratio(pd.Series([0.0]), pd.Series([0.0]))
     )
 
 
@@ -160,8 +165,8 @@ def test_seasons_share_one_colour_domain():
     summer = np.array([[519.0, 540.0], [530.0, 554.0]])
     valid = np.ones_like(winter, dtype=bool)
 
-    w = solar.render_scale("winter", winter, valid, summer, valid)
-    s = solar.render_scale("summer", summer, valid, winter, valid)
+    w = overlays_mod.render_scale("winter", winter, valid, summer, valid)
+    s = overlays_mod.render_scale("summer", summer, valid, winter, valid)
 
     assert (w["min"], w["max"]) == (s["min"], s["max"])
     assert w["min"] == 180.0 and w["max"] == 554.0
@@ -199,7 +204,7 @@ def test_a_flat_season_stays_flat_on_the_shared_domain():
     summer = np.array([[519.0, 530.0, 545.0, 554.0]])
     valid = np.ones_like(summer, dtype=bool)
 
-    shared = solar.render_scale("summer", summer, valid, winter, valid)
+    shared = overlays_mod.render_scale("summer", summer, valid, winter, valid)
     summer_shared = _ramp_fraction(summer, shared)
     winter_shared = _ramp_fraction(winter, shared)
     assert summer_shared < 0.15, "the flat season stays in a narrow slice"
@@ -207,14 +212,14 @@ def test_a_flat_season_stays_flat_on_the_shared_domain():
         "on one domain the structured season occupies far more of the ramp"
     )
 
-    own = solar.render_scale("annual", summer, valid)
+    own = overlays_mod.render_scale("annual", summer, valid)
     assert _ramp_fraction(summer, own) == 1.0, (
         "self-normalising spends the whole ramp on the flat season, "
         "which is the defect"
     )
 
     # and the colours follow: winter moves further than summer on one domain
-    travel = lambda a: _colour_travel(solar.terrain_rgba(
+    travel = lambda a: _colour_travel(overlays_mod.terrain_rgba(
         a, valid, shared["min"], shared["max"], shared["palette"]))
     assert travel(winter) > travel(summer)
 
@@ -225,7 +230,7 @@ def test_anisotropy_domain_keeps_the_parity_reference_visible():
     same irradiation. A percentile stretch of observed values, which never reach
     one, would drop that reference off the ramp.
     """
-    scale = solar.render_scale("anisotropy")
+    scale = overlays_mod.render_scale("anisotropy")
     assert scale["basis"] == "fixed"
     assert scale["reference"] == 1.0
     assert scale["min"] < 1.0 < scale["max"]
@@ -233,7 +238,7 @@ def test_anisotropy_domain_keeps_the_parity_reference_visible():
 
     observed = np.array([[0.33, 0.57, 0.83]])
     valid = np.ones_like(observed, dtype=bool)
-    rgba = solar.terrain_rgba(
+    rgba = overlays_mod.terrain_rgba(
         observed, valid, scale["min"], scale["max"], scale["palette"]
     )
     assert rgba[..., 3].all(), "every valid pixel is opaque"
@@ -247,10 +252,10 @@ def test_irradiation_does_not_use_the_vegetation_ramp():
     """
     import composite as comp
 
-    assert solar.PALETTE_IRRADIATION == "inferno"
-    assert solar.PALETTE_SHADING == "viridis"
-    for name in (solar.PALETTE_IRRADIATION, solar.PALETTE_SHADING,
-                 solar.PALETTE_ANISOTROPY):
+    assert overlays_mod.PALETTE_IRRADIATION == "inferno"
+    assert overlays_mod.PALETTE_SHADING == "viridis"
+    for name in (overlays_mod.PALETTE_IRRADIATION, overlays_mod.PALETTE_SHADING,
+                 overlays_mod.PALETTE_ANISOTROPY):
         assert name in comp.CONTINUOUS_STOPS
         assert comp.CONTINUOUS_STOPS[name] is not comp._RDYLGN
 
@@ -259,7 +264,7 @@ def test_degenerate_domain_does_not_divide_by_zero():
     """A constant layer still has to render."""
     flat = np.full((3, 3), 42.0)
     valid = np.ones_like(flat, dtype=bool)
-    rgba = solar.terrain_rgba(flat, valid, 42.0, 42.0, "inferno")
+    rgba = overlays_mod.terrain_rgba(flat, valid, 42.0, 42.0, "inferno")
     assert rgba.shape == (3, 3, 4)
     assert np.isfinite(rgba).all()
 
@@ -288,7 +293,7 @@ def _pit(h=16, w=16, height=200.0, floor=4):
 
 def test_horizon_sees_a_ridge_to_the_north():
     z = _ridge()
-    horizon, az = solar.horizon_angles(z, 30.0, 30.0)
+    horizon, az = poa_mod.horizon_angles(z, 30.0, 30.0)
     north = int(np.argmin(np.abs(az - 0.0)))
     south = int(np.argmin(np.abs(az - 180.0)))
     # a pixel just south of the wall looks up at it, and sees nothing behind
@@ -297,7 +302,7 @@ def test_horizon_sees_a_ridge_to_the_north():
 
 
 def test_flat_ground_has_no_horizon():
-    horizon, _ = solar.horizon_angles(np.zeros((16, 16)), 30.0, 30.0)
+    horizon, _ = poa_mod.horizon_angles(np.zeros((16, 16)), 30.0, 30.0)
     assert horizon.max() == 0.0
 
 
@@ -318,12 +323,12 @@ def test_shading_loss_is_zero_without_relief_and_positive_with_it():
     hist, edges = sun_position.beam_energy_histogram(df, solpos)
     assert hist.sum() > 0
 
-    flat = solar.shading_loss_fraction(
-        solar.horizon_angles(np.zeros((16, 16)), 30.0, 30.0)[0], hist, edges)
+    flat = poa_mod.shading_loss_fraction(
+        poa_mod.horizon_angles(np.zeros((16, 16)), 30.0, 30.0)[0], hist, edges)
     assert flat.max() == 0.0
 
-    walled = solar.shading_loss_fraction(
-        solar.horizon_angles(_pit(), 30.0, 30.0)[0], hist, edges)
+    walled = poa_mod.shading_loss_fraction(
+        poa_mod.horizon_angles(_pit(), 30.0, 30.0)[0], hist, edges)
     assert walled.max() > 0.0
     assert walled.max() <= 1.0
     assert walled[8, 8] > 0.0, "the pit floor loses beam energy"
@@ -348,12 +353,12 @@ def test_sky_view_factor_is_the_closed_form_for_a_uniform_horizon():
     a ninety test and be wrong everywhere between.
     """
     for deg in (0.0, 10.0, 30.0, 45.0, 60.0, 90.0):
-        got = solar.sky_view_factor(_uniform_horizon(deg))[0, 0]
+        got = poa_mod.sky_view_factor(_uniform_horizon(deg))[0, 0]
         assert abs(got - np.cos(np.radians(deg)) ** 2) < 1e-6, deg
-    assert solar.sky_view_factor(_uniform_horizon(0.0))[0, 0] == 1.0
+    assert poa_mod.sky_view_factor(_uniform_horizon(0.0))[0, 0] == 1.0
     # Not an exact zero: cos(pi/2) is 6.1e-17 in floating point, so a sky walled
     # to the zenith lands at 1.9e-15 rather than at 0.
-    assert solar.sky_view_factor(_uniform_horizon(90.0))[0, 0] < 1e-12
+    assert poa_mod.sky_view_factor(_uniform_horizon(90.0))[0, 0] < 1e-12
 
 
 def test_sky_view_factor_averages_over_sectors_not_within_them():
@@ -361,14 +366,14 @@ def test_sky_view_factor_averages_over_sectors_not_within_them():
     mixed = np.zeros((1, 1, 36), dtype=np.float32)
     mixed[:, :, :18] = 45.0
     want = 0.5 * np.cos(np.radians(45.0)) ** 2 + 0.5
-    assert abs(solar.sky_view_factor(mixed)[0, 0] - want) < 1e-6
+    assert abs(poa_mod.sky_view_factor(mixed)[0, 0] - want) < 1e-6
 
 
 def test_diffuse_loss_is_the_complement_and_stays_bounded():
-    loss = solar.diffuse_loss_fraction(_uniform_horizon(30.0))
+    loss = poa_mod.diffuse_loss_fraction(_uniform_horizon(30.0))
     assert abs(loss[0, 0] - 0.25) < 1e-6
-    assert solar.diffuse_loss_fraction(_uniform_horizon(0.0)).max() == 0.0
-    assert solar.diffuse_loss_fraction(_uniform_horizon(90.0)).max() == 1.0
+    assert poa_mod.diffuse_loss_fraction(_uniform_horizon(0.0)).max() == 0.0
+    assert poa_mod.diffuse_loss_fraction(_uniform_horizon(90.0)).max() == 1.0
 
 
 def test_diffuse_loss_reproduces_the_measured_valley_and_plain():
@@ -377,8 +382,8 @@ def test_diffuse_loss_reproduces_the_measured_valley_and_plain():
     and -0.04 percent on a plain. Those are the magnitudes the implementation
     has to land on, or it is measuring something else.
     """
-    valley = solar.diffuse_loss_fraction(_uniform_horizon(9.7))[0, 0]
-    plain = solar.diffuse_loss_fraction(_uniform_horizon(1.0))[0, 0]
+    valley = poa_mod.diffuse_loss_fraction(_uniform_horizon(9.7))[0, 0]
+    plain = poa_mod.diffuse_loss_fraction(_uniform_horizon(1.0))[0, 0]
     assert 0.025 < valley < 0.032, valley
     assert plain < 0.001, plain
 
@@ -388,23 +393,23 @@ def test_enclosure_gates_on_the_horizon_it_reports():
     The verdict travels with the evidence: a caller that applies the loss has to
     be able to print the horizon and the threshold it was judged against.
     """
-    below = solar.horizon_enclosure(_uniform_horizon(1.9))
-    at = solar.horizon_enclosure(_uniform_horizon(2.0))
+    below = poa_mod.horizon_enclosure(_uniform_horizon(1.9))
+    at = poa_mod.horizon_enclosure(_uniform_horizon(2.0))
     assert below["encloses"] is False
     assert at["encloses"] is True
-    assert at["threshold_deg"] == solar.SVF_MIN_MEAN_HORIZON_DEG
+    assert at["threshold_deg"] == poa_mod.SVF_MIN_MEAN_HORIZON_DEG
     assert abs(at["mean_horizon_deg"] - 2.0) < 1e-6
     # The threshold is where the loss is still under the rounding of every
     # figure this module publishes.
-    assert solar.diffuse_loss_fraction(
-        _uniform_horizon(solar.SVF_MIN_MEAN_HORIZON_DEG)
+    assert poa_mod.diffuse_loss_fraction(
+        _uniform_horizon(poa_mod.SVF_MIN_MEAN_HORIZON_DEG)
     ).max() < 0.002
 
 
 def test_sky_view_factor_of_an_absent_horizon_is_open_sky():
     """No horizon traced must read as nothing blocking, not as everything."""
-    assert solar.sky_view_factor(np.zeros((2, 2, 0), dtype=np.float32))[0, 0] == 1.0
-    assert solar.horizon_enclosure(np.zeros((2, 2, 0)))["encloses"] is False
+    assert poa_mod.sky_view_factor(np.zeros((2, 2, 0), dtype=np.float32))[0, 0] == 1.0
+    assert poa_mod.horizon_enclosure(np.zeros((2, 2, 0)))["encloses"] is False
 
 
 def test_diffuse_incidence_correction_is_applied_and_lowers_the_yield():
@@ -414,11 +419,11 @@ def test_diffuse_incidence_correction_is_applied_and_lowers_the_yield():
     strongly corrected term; the omission overstated the yield by ~0.75 percent.
     """
     import pvlib
-    sky = pvlib.iam.marion_diffuse("ashrae", 20.0, b=solar.IAM_ASHRAE_B)
+    sky = pvlib.iam.marion_diffuse("ashrae", 20.0, b=pv_mod.IAM_ASHRAE_B)
     assert sky["ground"] < sky["sky"] < 1.0
     assert 0.9 < sky["sky"] < 1.0
     # The beam relation and the diffuse one are the same coefficient.
-    assert solar.IAM_ASHRAE_B == 0.05
+    assert pv_mod.IAM_ASHRAE_B == 0.05
 
 
 def _three_years():
@@ -471,10 +476,10 @@ def test_the_window_is_not_the_named_season_helper():
     silêncio, que é exatamente o que aconteceu ao escrever esta."""
     idx = _three_years()
     assert sun_record.doy_window_mask(idx, "2026-02-19", 21).sum() != \
-        solar.season_mask(idx, "summer").sum() if "summer" in solar.SEASONS \
+        seasons_mod.season_mask(idx, "summer").sum() if "summer" in seasons_mod.SEASONS \
         else True
     assert sun_record.doy_window_mask.__code__.co_argcount == 3
-    assert solar.season_mask.__code__.co_argcount == 2
+    assert seasons_mod.season_mask.__code__.co_argcount == 2
 
 
 def _clear_day(hours=24, lat=-4.5, lon=-42.5):
